+++ resources/assets/src/js/admin.ts
@@ -11,6 +11,7 @@ import aside from './includes/aside';
 import autocomplete from './includes/autocomplete';
 import busy from './includes/busy';
 import collapse from './includes/collapse';
+import colorScheme from './includes/colorScheme';
 import dropdown from './includes/dropdown';
 import filter from './includes/filter';
 import {closeModal, createModal} from './includes/modals';
@@ -37,6 +38,7 @@ onLoad(($container) => {
     queryAll('[data-autocomplete]', autocomplete);
     queryAll('[data-busy]', busy);
     queryAll('[data-collapse]', collapse);
+    queryAll('[data-color-scheme]', colorScheme);
     queryAll('[popovertarget]', dropdown);
     queryAll('[data-filter]', filter);
     queryAll('[data-modal]', closeModal);