+++ resources/views/admin/layouts/main.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
  */
 
 use Hirtz\Skeleton\Assets\AdminAssetBundle;
+use Hirtz\Skeleton\Modules\Admin\Module;
 use Hirtz\Skeleton\Modules\Admin\TimezoneModal;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\AsideMenu;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\NavBar;
@@ -18,10 +19,12 @@ use Hirtz\Skeleton\Widgets\Navs\Breadcrumbs;
 use yii\helpers\Html;
 
 AdminAssetBundle::register($this);
+
+$colorScheme = Module::current()->getColorScheme();
 ?>
 <?php $this->beginPage() ?>
     <!DOCTYPE html>
-    <html lang="<?= Yii::$app->getI18n()->getLanguageCode(); ?>">
+    <html lang="<?= Yii::$app->getI18n()->getLanguageCode(); ?>"<?= $colorScheme ? ' data-theme="' . $colorScheme . '"' : ''; ?>>
     <head>
         <meta charset="<?= Yii::$app->charset ?>">
         <meta name="viewport" content="width=device-width,height=device-height,initial-scale=1">